refactor: fold DuploConfig into the config resource; controller loads it as a service
