+++ src/tests/test_config_resource.py
@@ -5,9 +5,8 @@ import pytest
 import yaml
 from unittest.mock import MagicMock
 
-from duplocloud.config import DuploConfig, VALID_CONTEXT_KEYS
 from duplocloud.errors import DuploError, DuploNotFound
-from duplo_resource.config import DuploConfigResource
+from duplo_resource.config import DuploConfigResource, VALID_CONTEXT_KEYS
 
 CONFIG_FIXTURE = {
   "current-context": "primary",
@@ -37,14 +36,19 @@ def write_fixture(tmp_path):
   return path
 
 
-def make_resource(tmp_path, ctx=None, missing=False):
-  """Build a config resource around a MagicMock duplo and a real store."""
-  path = str(tmp_path / "config") if missing else write_fixture(tmp_path)
+def make_resource_at(path, ctx=None):
+  """Build a config resource around a MagicMock duplo at a given path."""
   mock_duplo = MagicMock()
-  mock_duplo.config_store = DuploConfig(path)
+  mock_duplo.config_file = path
   mock_duplo._context = ctx
   mock_duplo.validate = False
   return DuploConfigResource(mock_duplo)
+
+
+def make_resource(tmp_path, ctx=None, missing=False):
+  """Build a config resource over the fixture file (or a missing one)."""
+  path = str(tmp_path / "config") if missing else write_fixture(tmp_path)
+  return make_resource_at(path, ctx)
 
 
 @pytest.mark.unit
@@ -57,7 +61,7 @@ class TestConfigResource:
     assert doc["contexts"][0]["helpdesk_token"] == "REDACTED"
     assert doc["config"] == {"browser": "firefox"}
     # the cached document must keep the real values
-    assert resource.store.data["contexts"][0]["token"] == "secret-token"
+    assert resource.data["contexts"][0]["token"] == "secret-token"
 
   def test_missing_file_errors(self, tmp_path):
     """Test view/get/unset/use raise a 500 when the file is missing."""
@@ -81,20 +85,20 @@ class TestConfigResource:
   def test_invalid_keys_rejected(self, tmp_path):
     """Test get/set/unset reject keys outside the allowlist."""
     resource = make_resource(tmp_path)
-    before = open(resource.store.path).read()
+    before = open(resource.path).read()
     for call in (lambda: resource.get("bogus"),
                  lambda: resource.set("bogus", "x"),
                  lambda: resource.unset("name")):
       with pytest.raises(DuploError, match="Valid keys:"):
         call()
-    assert open(resource.store.path).read() == before
+    assert open(resource.path).read() == before
 
   def test_set_writes_and_preserves_unknown_keys(self, tmp_path):
     """Test set rewrites the file preserving unrelated content."""
     resource = make_resource(tmp_path)
     result = resource.set("workspace", "my-ws")
     assert result == {"message": "set workspace in context 'primary'"}
-    with open(resource.store.path) as f:
+    with open(resource.path) as f:
       doc = yaml.safe_load(f)
     primary = doc["contexts"][0]
     assert primary["workspace"] == "my-ws"
@@ -108,7 +112,7 @@ class TestConfigResource:
     """Test boolean keys coerce true/false and reject other values."""
     resource = make_resource(tmp_path)
     resource.set("interactive", "True")
-    with open(resource.store.path) as f:
+    with open(resource.path) as f:
       doc = yaml.safe_load(f)
     assert doc["contexts"][0]["interactive"] is True
     with pytest.raises(DuploError, match="use true or false"):
@@ -118,7 +122,7 @@ class TestConfigResource:
     """Test set with --ctx creates a missing context in the file."""
     resource = make_resource(tmp_path, ctx="newctx")
     resource.set("host", "https://new.duplocloud.net")
-    with open(resource.store.path) as f:
+    with open(resource.path) as f:
       doc = yaml.safe_load(f)
     names = [c["name"] for c in doc["contexts"]]
     assert "newctx" in names
@@ -129,7 +133,7 @@ class TestConfigResource:
     """Test set with --ctx scaffolds a brand new config file."""
     resource = make_resource(tmp_path, ctx="newctx", missing=True)
     resource.set("host", "https://new.duplocloud.net")
-    with open(resource.store.path) as f:
+    with open(resource.path) as f:
       doc = yaml.safe_load(f)
     assert doc["current-context"] == "newctx"
     assert doc["contexts"] == [
@@ -140,14 +144,14 @@ class TestConfigResource:
     resource = make_resource(tmp_path, missing=True)
     with pytest.raises(DuploError, match="pass --ctx"):
       resource.set("host", "https://x")
-    assert not os.path.exists(resource.store.path)
+    assert not os.path.exists(resource.path)
 
   def test_unset_removes_key_and_absent_is_noop(self, tmp_path):
     """Test unset removes a set key and succeeds on an absent one."""
     resource = make_resource(tmp_path)
     result = resource.unset("tenant")
     assert result == {"message": "unset tenant in context 'primary'"}
-    with open(resource.store.path) as f:
+    with open(resource.path) as f:
       doc = yaml.safe_load(f)
     assert "tenant" not in doc["contexts"][0]
     assert resource.unset("workspace") == {
@@ -158,12 +162,12 @@ class TestConfigResource:
     resource = make_resource(tmp_path)
     assert resource.use("other") == {
       "message": "switched to context 'other'"}
-    with open(resource.store.path) as f:
+    with open(resource.path) as f:
       assert yaml.safe_load(f)["current-context"] == "other"
-    before = open(resource.store.path).read()
+    before = open(resource.path).read()
     with pytest.raises(DuploError, match="not found in config"):
       resource.use("nope")
-    assert open(resource.store.path).read() == before
+    assert open(resource.path).read() == before
     with pytest.raises(DuploError, match="context name is required"):
       resource.use()
 
@@ -175,30 +179,50 @@ class TestConfigResource:
     assert resource("use", "other") == {
       "message": "switched to context 'other'"}
 
-  def test_store_stays_coherent_after_write(self, tmp_path):
-    """Test the shared store reflects writes without a re-read."""
+  def test_cache_stays_coherent_after_write(self, tmp_path):
+    """Test the cached document reflects writes without a re-read."""
     resource = make_resource(tmp_path)
     resource.set("workspace", "my-ws")
-    assert resource.duplo.config_store.data[
-      "contexts"][0]["workspace"] == "my-ws"
+    assert resource.data["contexts"][0]["workspace"] == "my-ws"
 
 
 @pytest.mark.unit
-class TestDuploConfig:
+class TestControllerConfigService:
+  """The controller loads the config resource as its config service."""
+
+  def test_settings_and_context_read_through_service(self, tmp_path):
+    from duplocloud.controller import DuploCtl
+    duplo = DuploCtl(config_file=write_fixture(tmp_path))
+    assert isinstance(duplo.config_svc, DuploConfigResource)
+    # the service instance is cached, so its parsed-document cache is
+    # shared between the controller read path and writers
+    assert duplo.config_svc is duplo.config_svc
+    assert duplo.settings["current-context"] == "primary"
+    assert duplo.context["host"] == "https://primary.duplocloud.net"
+
+  def test_write_through_service_updates_settings(self, tmp_path):
+    from duplocloud.controller import DuploCtl
+    duplo = DuploCtl(config_file=write_fixture(tmp_path))
+    duplo.config_svc.set("workspace", "my-ws")
+    assert duplo.settings["contexts"][0]["workspace"] == "my-ws"
+
+
+@pytest.mark.unit
+class TestConfigStore:
   def test_missing_file_errors(self, tmp_path):
     """Test reading a missing config file raises a 500."""
-    store = DuploConfig(str(tmp_path / "config"))
+    store = make_resource_at(str(tmp_path / "config"))
     with pytest.raises(DuploError, match="Duplo config not found"):
       store.data
 
   def test_get_context_lookup_and_errors(self, tmp_path):
     """Test context lookup by name plus both error messages."""
-    store = DuploConfig(write_fixture(tmp_path))
+    store = make_resource_at(write_fixture(tmp_path))
     assert store.get_context()["name"] == "primary"
     assert store.get_context("other")["name"] == "other"
     with pytest.raises(DuploError, match="Portal 'nope' not found"):
       store.get_context("nope")
-    empty = DuploConfig(str(tmp_path / "empty"))
+    empty = make_resource_at(str(tmp_path / "empty"))
     with open(empty.path, "w") as f:
       yaml.safe_dump({"contexts": []}, f)
     with pytest.raises(DuploError, match="context not set"):
@@ -206,7 +230,7 @@ class TestDuploConfig:
 
   def test_missing_contexts_key_is_not_a_keyerror(self, tmp_path):
     """Test a config with no contexts key raises a DuploError."""
-    store = DuploConfig(str(tmp_path / "config"))
+    store = make_resource_at(str(tmp_path / "config"))
     with open(store.path, "w") as f:
       yaml.safe_dump({"current-context": "primary"}, f)
     with pytest.raises(DuploError, match="Portal 'primary' not found"):
@@ -215,7 +239,7 @@ class TestDuploConfig:
   def test_save_is_atomic_and_private(self, tmp_path):
     """Test save creates parent dirs, chmods 0o600, leaves no tmp."""
     path = str(tmp_path / "deep" / "dir" / "config")
-    store = DuploConfig(path)
+    store = make_resource_at(path)
     store.save({"current-context": None, "contexts": []})
     mode = stat.S_IMODE(os.stat(path).st_mode)
     assert mode == 0o600
@@ -225,7 +249,7 @@ class TestDuploConfig:
 
   def test_empty_file_loads_as_scaffold(self, tmp_path):
     """Test an empty config file loads as a scaffold, not None."""
-    store = DuploConfig(str(tmp_path / "config"))
+    store = make_resource_at(str(tmp_path / "config"))
     with open(store.path, "w") as f:
       f.write("\n")
     assert store.data == {"current-context": None, "contexts": []}
@@ -239,9 +263,9 @@ class TestDuploConfig:
     A failed write must never orphan a temp file holding tokens,
     regardless of the process umask.
     """
-    import duplocloud.config as config_module
+    import duplo_resource.config as config_module
     path = str(tmp_path / "config")
-    store = DuploConfig(path)
+    store = make_resource_at(path)
     monkeypatch.setattr(
       config_module.yaml, "safe_dump",
       MagicMock(side_effect=RuntimeError("boom")))
@@ -260,7 +284,7 @@ class TestDuploConfig:
     fail the write instead of being followed or overwritten.
     """
     path = str(tmp_path / "config")
-    store = DuploConfig(path)
+    store = make_resource_at(path)
     tmp = f"{path}.tmp.{os.getpid()}"
     victim = str(tmp_path / "victim")
     with open(victim, "w") as f:
@@ -273,12 +297,12 @@ class TestDuploConfig:
 
   def test_non_mapping_config_errors(self, tmp_path):
     """Test list/scalar roots and malformed YAML raise DuploError."""
-    store = DuploConfig(str(tmp_path / "config"))
+    store = make_resource_at(str(tmp_path / "config"))
     with open(store.path, "w") as f:
       f.write("- just\n- a\n- list\n")
     with pytest.raises(DuploError, match="must be a YAML mapping"):
       store.data
-    broken = DuploConfig(str(tmp_path / "broken"))
+    broken = make_resource_at(str(tmp_path / "broken"))
     with open(broken.path, "w") as f:
       f.write("contexts: [unclosed\n")
     with pytest.raises(DuploError, match="Invalid YAML"):
@@ -287,8 +311,8 @@ class TestDuploConfig:
   def test_failed_save_leaves_cache_untouched(self, tmp_path,
                                               monkeypatch):
     """Test a failed write never leaves phantom values in the cache."""
-    import duplocloud.config as config_module
-    store = DuploConfig(write_fixture(tmp_path))
+    import duplo_resource.config as config_module
+    store = make_resource_at(write_fixture(tmp_path))
     assert store.get_context()["tenant"] == "dev01"
     monkeypatch.setattr(
       config_module, "_atomic_write_yaml",
@@ -316,14 +340,12 @@ class TestWorkspaceUse:
     ws = DuploWorkspace.__new__(DuploWorkspace)
     ws.duplo = MagicMock()
     ws.find = MagicMock(return_value={"name": "Canonical"})
-    config = MagicMock()
-    ws.duplo.load.return_value = config
-    config.set.return_value = {
+    ws.duplo.config_svc.set.return_value = {
       "message": "set workspace in context 'primary'"}
     result = DuploWorkspace.use(ws, "canonical")
     ws.find.assert_called_once_with("canonical")
-    ws.duplo.load.assert_called_once_with("config")
-    config.set.assert_called_once_with("workspace", "Canonical")
+    ws.duplo.config_svc.set.assert_called_once_with(
+      "workspace", "Canonical")
     assert result == {"message": "set workspace in context 'primary'"}
 
   def test_use_requires_name_and_propagates_not_found(self):
@@ -336,4 +358,4 @@ class TestWorkspaceUse:
     ws.find = MagicMock(side_effect=DuploNotFound("nope", "workspace"))
     with pytest.raises(DuploNotFound):
       DuploWorkspace.use(ws, "nope")
-    ws.duplo.load.assert_not_called()
+    ws.duplo.config_svc.set.assert_not_called()
